+++ tiger-common/src/main/java/de/gematik/test/tiger/common/data/config/AdditionalConfigurationFileProperty.java
@@ -16,17 +16,12 @@
 
 package de.gematik.test.tiger.common.data.config;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
 @Data
-public class AdditionalYamlProperty {
+public class AdditionalConfigurationFileProperty {
 
   private String filename;
   private String baseKey;
+  private ConfigurationFileType type = ConfigurationFileType.YAML;
 }